refactor(web): rename patterns-selection to patterns-form
Aligns with form namespace naming convention where forms live in
*-form directories. Also updates conventions.md documentation to use
proper ValidationResult and FieldsValidationResult types in examples.
